src/functions/requirements.txt
azure-functions==1.14.0
azure-identity==1.13.0
azure-keyvault==4.2.0
azure-storage-blob==12.16.0
azure-monitor-ingestion==1.0.1
